Formed all changes for the validation functions being moved to their own validation.js file.
