+++ client/src/pages/Dashboard/views/UserSettings/components/ClassroomModal.js
@@ -1,4 +1,5 @@
-import Form, { createValidator } from "components/Form";
+import Form from "components/Form";
+import { createValidator } from "utils/validation";
 import Pulse from "components/Pulse";
 import React, { useEffect, useState } from "react";
